Ensure references remain valid after source is moved
References &T to a type T are invalidated when T is moved, the same
applies to core::ptr::Unique, which is used internally by Vec. This
makes construction of a self-referential types such as TemplateBuf or
ByteTemplateBuf a safety hazard.

Introduce two helper types, aliasable::Bytes and aliasable::String,
that store the raw parts of the original vectors, such that shared
references can be obtained from the raw (non-unique) pointers.

The implementation is inspired by the aliasable crate, which is was
created for that very purpose.

See the following blog post by Hrvoje Nikšić for more details:
https://morestina.net/blog/1868/self-referential-types-for-fun-and-profit